requirements.txt
allure-python-commons
selenium==4.39.0
pytest==9.0.2
selene==2.0.0rc9
allure-pytest==2.15.3